Party + Party Leader Switching Functionality
Basically fully implemented Player Party functionality.

Created CharacterPartyReference class for Party functionality.

Created CharacterPartyInterface class for any needed Character-to-Party communication.

Created the CharacterInputManager class, of which all CharacterInputManagers (CharacterPlayerInputManager, CharacterPartyInputManager, and CharacterCOMInputManager) inherit.

Created the CharacterCameraManager class (and in turn, PartyCameraManager inherits from it).

A Player-controlled party is constituted by a Party node, a PartyCameraManager child, and 1 or more Character children with a CharacterPartyReference and CharacterPartyInterface children. This is the minimum required configuration for a Party.