FSM-ified Stage Interface, Inspect Interface + Accomp. Bug Fixes
I'm... so tired. Mission accomplished, though.

Renamed BehaviourToolkit's Blackboard class name to BTBlackboard to fix the ambiguity between it and Beehave's Blackboard class.

Added an InspectCamera Camera3D node to all Actors to give a proper camera angle for them when inspected.

Renamed the project input mappings. what was strafe_left and strafe_right is now left and right, and forward and backward is now up and down.

Actors are now rotated to face the opposite team.

Added clicking and holding functionality to StageCamera.

Added the HoldCircle, a visual for when the player is clicking and holding. Will always follow the mouse, but will be invisible without using set_progress().

Started blocking out the Inspect Interface, but it is currently unfinished. Highly unfinished.

Moved a majority of StageInterface functionality to a BehaviourToolkit FiniteStateMachine.