Added Mira and Winsey Models + Accomp. Bug Fixes
Fixed a bug where if the game was exited while in full screen, changing the window mode to windowed would not resize the window to the game's proper resolution of 1280 by 960.

All player party members now use their placeholder models.

When switching to a party member, pathfinding for the player-controlled character is now properly stopped. This fixes a bug where pathfinding calculations continued to be made even when the player has full control of the character.

Fixed the overworld and stage WorldEnvironment nodes as promised.