Inspect Interface Complete
Likely the last major commit of the month. I'm very happy with the progress made so far.

Added new ActorMetadata resource for providing an Actor a name, level (likely will be moved to an experience-based class) and description.

ActsManager class now has functions for getting act PackedScenes, mainly for use in the inspect interface.

Addeed StatusBarPanel and StatusEffectPanel for showing actor status in the inspect interface.

Properly using the InspectActor state on StageInterface.